pyproject.toml
[project]
name = "maim-web-backend"
version = "0.1.0"
description = "Web Backend for MaimBot"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "aiofiles",
    "httpx",
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
